refactor(ioxide): give the TCP listener its own partial, matching QUIC
IoxideServer.Quic.cs owned the QUIC listener while the TCP one was inlined in
StartAsync, so the two transports read as different kinds of thing when they are
the same kind. The TCP half now sits in IoxideServer.Tcp.cs behind WithTcp, the
mirror of WithQuic, and StartAsync says what it does:

    var serverConfig = WithTcp(BuildServerConfig());

    if (_quicRequested is { } quicEndPoint)
    {
        serverConfig = WithQuic(serverConfig, quicEndPoint);
    }

BuildServerConfig keeps only what is not a listener - the reactors, and DualStack,
which applies to the TCP listener and the UDP socket alike.

IoxideServer.Tls.cs becomes IoxideServer.Tcp.Tls.cs: it configures OpenSSL, which
terminates the TCP protocols only, and QUIC's TLS is ngtcp2's business in the
other file. RequiresClientCertificate is the one thing both ask, and now says so.
